A slight re-focussing of the addSources function - renamed to processSources and added a flag itsAddSources that means the actual addition can be switched off. Have also added a new executable convertPositions, that just does the precession of source positions.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@2438 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/simulations/trunk/simulationutilities/HIprofileS3SAX.cc
@@ -110,7 +110,7 @@ namespace askap {
             /// to set up the profile description.
             /// @param line A line from the ascii input file
 
-            float maj, min, pa, alpha, beta;
+            double maj, min, pa, alpha, beta;
             std::stringstream ss(line);
             ss >> this->itsRA >> this->itsDec >> this->itsIntFlux >> alpha >> beta >> maj >> min >> pa >> this->itsRedshift >> this->itsMHI
             >> this->itsFlux0 >> this->itsFluxPeak >> this->itsWidthPeak >> this->itsWidth50 >> this->itsWidth20;
